+++ code/modules/resleeving/computers.dm
@@ -1,12 +1,13 @@
 /obj/machinery/computer/transhuman/resleeving
 	name = "resleeving control console"
+	desc = "A control console for the resleeving and grower pods. It has a small slot for inserting a mirror tool for direct transfer of mirrors."
 	catalogue_data = list(///datum/category_item/catalogue/information/organization/vey_med,
 						/datum/category_item/catalogue/technology/resleeving)
 	icon_keyboard = "med_key"
 	icon_screen = "dna"
 	light_color = "#315ab4"
 	circuit = /obj/item/circuitboard/resleeving_control
-	req_access = list(access_heads) //Only used for record deletion right now.
+	req_access = list(ACCESS_COMMAND_BRIDGE) //Only used for record deletion right now.
 	var/list/pods = list() //Linked grower pods.
 	var/list/spods = list()
 	var/list/sleevers = list() //Linked resleeving booths.
@@ -99,6 +100,7 @@
 			hasmirror = MT.imp
 			MT.imp = null
 			user.visible_message("Mirror successfully transferred.")
+			MT.update_icon()
 		else
 			if(!MT.imp)
 				user.visible_message("This Mirror Installation Tool is empty.")
@@ -109,6 +111,7 @@
 			hasmirror = MT.imp
 			MT.imp = null
 			user.visible_message("Mirror successfully transferred.")
+			MT.update_icon()
 		else
 			if(!MT.imp)
 				user.visible_message("This Mirror Installation Tool is empty.")
@@ -123,7 +126,7 @@
 
 	if(hasmirror)
 		to_chat(usr, "You eject the mirror.")
-		usr.put_in_hands(hasmirror)
+		usr.put_in_hands_or_drop(hasmirror)
 		hasmirror = null
 		active_mr = null
 	else
@@ -132,7 +135,7 @@
 /obj/machinery/computer/transhuman/resleeving/attack_ai(mob/user as mob)
 	return attack_hand(user)
 
-/obj/machinery/computer/transhuman/resleeving/attack_hand(mob/user as mob)
+/obj/machinery/computer/transhuman/resleeving/attack_hand(mob/user, list/params)
 	user.set_machine(src)
 	add_fingerprint(user)
 
@@ -428,12 +431,15 @@
 	icon_state = "cmoemergency"
 	item_state = "card-id"
 
-/obj/item/cmo_disk_holder/attack_self(var/mob/attacker)
+/obj/item/cmo_disk_holder/attack_self(mob/user)
+	. = ..()
+	if(.)
+		return
 	playsound(src, 'sound/items/poster_ripped.ogg', 50)
-	to_chat(attacker, "<span class='warning'>You tear open \the [name].</span>")
-	attacker.temporarily_remove_from_inventory(src, INV_OP_FORCE | INV_OP_SHOULD_NOT_INTERCEPT | INV_OP_SILENT)
+	to_chat(user, "<span class='warning'>You tear open \the [name].</span>")
+	user.temporarily_remove_from_inventory(src, INV_OP_FORCE | INV_OP_SHOULD_NOT_INTERCEPT | INV_OP_SILENT)
 	var/obj/item/disk/transcore/newdisk = new(get_turf(src))
-	attacker.put_in_hands(newdisk)
+	user.put_in_hands_or_drop(newdisk)
 	qdel(src)
 
 /obj/item/disk/transcore
