+++ code/modules/xenobio/items/slime_objects.dm
@@ -7,7 +7,10 @@
 	description_info = "Use in your hand to attempt to create a Promethean.  It functions similarly to a positronic brain, in that a ghost is needed to become the Promethean."
 	var/searching = 0
 
-/obj/item/slime_cube/attack_self(mob/user as mob)
+/obj/item/slime_cube/attack_self(mob/user)
+	. = ..()
+	if(.)
+		return
 	if(!searching)
 		to_chat(user, "<span class='warning'>You stare at the slimy cube, watching as some activity occurs.</span>")
 		icon_state = "slime cube active"
@@ -78,14 +81,22 @@
 	icon_state = "slime_crystal_small"
 	w_class = ITEMSIZE_TINY
 	origin_tech = list(TECH_MAGNET = 6, TECH_BLUESPACE = 3)
-	force = 1 //Needs a token force to ensure you can attack because for some reason you can't attack with 0 force things
+	damage_force = 1 //Needs a token damage_force to ensure you can attack because for some reason you can't attack with 0 damage_force things
 
-/obj/item/slime_crystal/apply_hit_effect(mob/living/target, mob/living/user, var/hit_zone)
-	target.visible_message("<span class='warning'>\The [target] has been teleported with \the [src] by \the [user]!</span>")
-	safe_blink(target, 14)
+/obj/item/slime_crystal/melee_mob_hit(mob/target, mob/user, clickchain_flags, list/params, mult, target_zone, intent)
+	. = ..()
+	var/mob/living/L = target
+	if(!istype(L))
+		return
+	L.visible_message("<span class='warning'>\The [L] has been teleported with \the [src] by \the [user]!</span>")
+	safe_blink(L, 14)
 	qdel(src)
+	return . | CLICKCHAIN_DO_NOT_PROPAGATE
 
 /obj/item/slime_crystal/attack_self(mob/user)
+	. = ..()
+	if(.)
+		return
 	user.visible_message("<span class='warning'>\The [user] teleports themselves with \the [src]!</span>")
 	safe_blink(user, 14)
 	qdel(src)
